refactor: Move each utility to respective packages
