+++ src/mg/emberframework/util/http/UrlParser.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util;
+package mg.emberframework.util.http;
 
 import java.net.URI;
 import java.net.URISyntaxException;
