+++ src/mg/emberframework/util/reflection/ClassUtils.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util;
+package mg.emberframework.util.reflection;
 
 import java.io.InputStream;
 import java.io.OutputStream;
