+++ src/mg/emberframework/util/reflection/ReflectUtils.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util;
+package mg.emberframework.util.reflection;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
